requirements.txt
pillow==10.0.0
pyaudio==0.2.14
pyautogui==0.9.54
customtkinter==5.2.0
numpy==1.24.3
playsound==1.3.0
pynput==1.7.7
pyperclip==1.9.0
requests==2.32.3
scipy==1.10.1
sounddevice==0.4.6
